Move option listeners to core
